+++ src/main/java/io/jenkins/plugins/sample/helper/PluginHelper.java
@@ -1,4 +1,4 @@
-package io.jenkins.plugins.sample.util;
+package io.jenkins.plugins.sample.helper;
 
 import java.io.IOException;
 import java.net.HttpURLConnection;
@@ -8,6 +8,9 @@ import java.util.Base64;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
+/**
+ * A helper class having static method for common functions which will be use across plugin
+ */
 public class PluginHelper {
 
     public static int httpPostBasicAuth(String url, String username, String password)
@@ -23,9 +26,19 @@ public class PluginHelper {
         return connection.getResponseCode();
     }
 
-    public static boolean isInvalidField(String field, String regex) {
+    /**
+     * Checks if the value provided matches the regex pattern provided
+     * @param value
+     *        the value to check for against teh regex
+     * @param regex
+     *        the regex against which to validate value
+     * @return
+     *       boolean true if value matches regex
+     *               false if value doesn't match the regex
+     */
+    public static boolean isMatch(String value, String regex) {
         Pattern pattern = Pattern.compile(regex);
-        Matcher matcher = pattern.matcher(field);
-        return !matcher.matches();
+        Matcher matcher = pattern.matcher(value);
+        return matcher.matches();
     }
 }
